Add parallelism.py
